+++ pymfy/api/devices/camera_protect.py
@@ -1,4 +1,4 @@
-from typing import Any
+from typing import cast
 
 from pymfy.api.devices.base import SomfyDevice
 from pymfy.api.model import Command
@@ -8,11 +8,11 @@ class CameraProtect(SomfyDevice):
     """Class to represent a camera"""
 
     def close_shutter(self) -> None:
-        self.send_command(Command("close"))
+        self.send_command(Command("shutter_close"))
 
     def open_shutter(self) -> None:
-        self.send_command(Command("open"))
+        self.send_command(Command("shutter_open"))
 
-    # TODO Waiting documentation to know what's the returned type
-    def get_shutter_position(self) -> Any:
-        return self.get_state("shutter_position")
+    def get_shutter_position(self) -> str:
+        """ Possible returned values are opened and closed """
+        return cast(str, self.get_state("shutter_position"))